# docs/swagger.yaml
definitions:
...
  RestApi_UnUpset_internal_delivery_dto.CreateTimerRequest:
    properties:
      duration:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: Продолжительность таймера (обязательное поле, минимум 1 наносекунда)
        minimum: 1
    required:
    - duration
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.Response:
    properties:
      data:
        description: Данные ответа (опционально)
      message:
        description: Сообщение (обычно используется для ошибок)
        type: string
      status:
        description: Статус ответа (success/error)
        type: string
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.TimerResponse:
    properties:
      duration:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: Продолжительность таймера в наносекундах
      id:
        description: Идентификатор таймера
        type: integer
      user_id:
        description: Идентификатор пользователя, создавшего таймер
        type: integer
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8080
info:
  contact: {}
  description: API для приложения UnUpset
  title: UnUpset API
  version: "1.0"
paths:
...
  /api/timers:
    get:
      description: Получение списка всех таймеров текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.TimerResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - UserAuth: []
      summary: Получение всех таймеров пользователя
      tags:
      - timers
    post:
      consumes:
      - application/json
      description: Создание нового таймера для текущего пользователя
      parameters:
      - description: Данные таймера
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.CreateTimerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.TimerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание нового таймера
      tags:
      - timers
  /api/timers/{id}:
    delete:
      description: Удаление существующего таймера (только своего)
      parameters:
      - description: ID таймера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - UserAuth: []
      summary: Удаление таймера
      tags:
      - timers
    get:
      description: Получение конкретного таймера по его ID
      parameters:
      - description: ID таймера
        format: uint
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.TimerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение таймера по ID
      tags:
      - timers
...
securityDefinitions:
  ApiKeyAuth:
    description: Для тестирования в Swagger UI (на макбуке логин работает только через
      хром, в сафари протестировать не получается)
    in: cookie
    name: mysession
    type: apiKey
swagger: "2.0"